requirements.txt
fastapi
uvicorn
python-dotenv
supabase
pydantic
sentence-transformers
faiss-cpu
pdfplumber
python-docx
numpy
scipy
transformers
torch
